requirements.txt
absl-py==2.1.0
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.2.0
protobuf==4.25.3
pyarrow==16.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
streamlit==1.34.0
streamlit-option-menu==0.3.12
sympy==1.12
tbb==2021.11.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
toml==0.10.2
toolz==0.12.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
